Move interfaces into the different packages
