--- packages/interfaces/.swcrc
@@ -1,31 +1,0 @@
-{
-  "jsc": {
-    "parser": {
-      "syntax": "typescript"
-    },
-    "target": "es2022",
-    "experimental": {
-      "plugins": [
-        [
-          "@swc/plugin-transform-imports",
-          {
-            "^(\\.{1,2}\\/.*?)$": {
-              "skipDefaultConversion": true,
-              "transform": "{{matches.[1]}}.js"
-            }
-          }
-        ]
-      ]
-    }
-  },
-  "module": {
-    "type": "es6"
-  },
-  "exclude": [
-    ".*\\.test\\.ts$",
-    "node_modules",
-    "dist",
-    "coverage",
-    "jest.config.*"
-  ]
-}
