--- packages/interfaces/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@positronic/interfaces",
-  "version": "0.0.1",
-  "description": "Service interfaces for Positronic",
-  "type": "module",
-  "main": "dist/src/index.js",
-  "types": "dist/types/index.ts",
-  "scripts": {
-    "tsc": "tsc --project tsconfig.json",
-    "swc": "swc src -d dist",
-    "build": "npm run tsc && npm run swc",
-    "clean": "rm -rf tsconfig.tsbuildinfo dist"
-  },
-  "dependencies": {
-    "zod": "^3.24.1"
-  }
-}
