--- packages/interfaces/src/index.ts
@@ -1,2 +1,0 @@
-export * from './resource-loader';
-export * from './shell';
