--- packages/interfaces/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "emitDeclarationOnly": true,
-    "declarationDir": "./dist/types"
-  },
-  "include": ["**/*"],
-  "exclude": ["dist", "node_modules", "**/*.test.ts"]
-}
