+++ packages/positronic/package.json
@@ -1,21 +1,33 @@
 {
   "name": "@positronic/core",
   "version": "0.0.1",
-  "description": "A DSL for AI Workflows",
+  "description": "Core services and tools for positronic AI workflows",
   "type": "module",
   "main": "dist/src/index.js",
-  "types": "dist/types/index.d.ts",
+  "types": "dist/types/index.ts",
   "scripts": {
     "tsc": "tsc --project tsconfig.json",
-    "build": "npm run tsc && swc src -d dist",
-    "clean": "rm -rf tsconfig.tsbuildinfo dist"
+    "swc": "swc src -d dist",
+    "build": "npm run tsc && npm run swc",
+    "clean": "rm -rf tsconfig.tsbuildinfo dist",
+    "test": "jest --silent"
   },
   "dependencies": {
+    "@positronic/resources": "^0.0.1",
+    "@positronic/shell": "^0.0.1",
+    "axios": "^1.6.7",
+    "chalk": "^5.3.0",
+    "commander": "^12.0.0",
     "fast-json-patch": "^3.1.1",
+    "inquirer": "^9.2.15",
+    "js-yaml": "^4.1.0",
+    "openai": "^4.29.0",
     "uuid": "^11.0.5",
     "zod": "^3.24.1"
   },
   "devDependencies": {
+    "@types/inquirer": "^9.0.7",
+    "@types/js-yaml": "^4.0.9",
     "@types/uuid": "^10.0.0"
   }
 }
