+++ packages/resources/package.json
@@ -13,7 +13,6 @@
   },
   "dependencies": {
     "@positronic/core": "^0.0.1",
-    "@positronic/interfaces": "^0.0.1",
     "zod": "^3.24.1"
   }
 }
