+++ packages/resources/src/types.ts
