+++ packages/shell/package.json
@@ -13,7 +13,6 @@
   },
   "dependencies": {
     "@positronic/core": "^0.0.1",
-    "@positronic/interfaces": "^0.0.1",
     "ssh2": "^1.14.0",
     "zod": "^3.24.1"
   }
